+++ ClassesExt/TestExt.py
@@ -1,0 +1,4 @@
+# class inside the module
+class TestExt:
+    def Comprendre():
+        print('Salut je suis le test')
