+++ ModulesCoeur/Initialisation.py
@@ -1,5 +1,4 @@
 import os
-from os import walk
 from ClassesCoeur.Utilisateur import *
 from ClassesCoeur.ImportClassesExt import *
 
@@ -17,19 +16,3 @@ def initialiser_infos_user(USERInterface, USER_NAME):
     nom_user = USERInterface.demande_utilisateur('Comment dois-je vous appeller ?')
     user = Utilisateur(nom_user, 'AZERTY', USER_NAME)
     USERInterface.informer_utilisateur('Félicitations, votre compte est initialisé !')
-
-
-def initialiser_modules_EXT():
-    # Cherche tous les fichiers du répertoire
-    for (repertoire, sousRepertoires, liste_fichiers) in walk('./ClassesExt/'):
-        # Se limite au répertoire des Classes Externes
-        if repertoire == './ClassesExt/':
-            # Liste les fichiers
-            for fichier in liste_fichiers:
-                # Suppression de l'extension
-                fichier = fichier.replace('.py', '')
-                if fichier != 'Comprendre':
-                    # Création de l'objet
-                    obj = ImportClassesExt(fichier, fichier)
-                    print(fichier)
-
